De-duplicate resource value decoding
