+++ ApkReader/AxmlApi.cs
@@ -36,27 +36,6 @@ namespace AlphaOmega.Debug
 			Declaration = 0x00100110,
 		}
 
-		/// <summary>Attibute value type</summary>
-		public enum AttributeValue
-		{
-			NULL = 0x00,
-			REFERENCE = 0x01,
-			ATTRIBUTE = 0x02,
-			STRING = 0x03,
-			FLOAT = 0x04,
-			DIMENSION = 0x05,
-			FRACTION = 0x06,
-
-			INT_DEC = 0x10,
-			INT_HEX = 0x11,
-			INT_BOOLEAN = 0x12,
-
-			INT_COLOR_ARGB8 = 0x1c,
-			INT_COLOR_RGB8 = 0x1d,
-			INT_COLOR_ARGB4 = 0x1e,
-			INT_COLOR_RGB4 = 0x1f,
-		}
-
 		/// <summary>AXML header file</summary>
 		[StructLayout(LayoutKind.Sequential)]
 		public struct AxmlHeader
@@ -181,123 +160,7 @@ namespace AlphaOmega.Debug
 			public Int32 value;
 
 			/// <summary>Value type</summary>
-			public AttributeValue ValueType => (AttributeValue)(this._valueType >> 24);/*Padding +3 bytes*/
-
-			/// <summary>Complex data: bit location of unit information.</summary>
-			private const Int32 COMPLEX_UNIT_SHIFT = 0;
-
-			/// <summary>
-			/// Complex data: mask to extract unit information (after shifting by <see cref="COMPLEX_UNIT_SHIFT"/>.
-			/// This gives us 16 possible types, as defined below.
-			/// </summary>
-			private const Int32 COMPLEX_UNIT_MASK = 0xf;
-
-			/// <summary>Complex data: bit location of mantissa information.</summary>
-			private const Int32 COMPLEX_MANTISSA_SHIFT = 8;
-
-			/// <summary>
-			/// Complex data: mask to extract mantissa information (after shifting by <see cref="COMPLEX_MANTISSA_SHIFT"/>.
-			/// This gives us 23 bits of precision; the top bit is the sign.
-			/// </summary>
-			private const Int32 COMPLEX_MANTISSA_MASK = 0xffffff;
-
-			/// <summary>Complex data: where the radix information is, telling where the decimal place appears in the mantissa.</summary>
-			private const Int32 COMPLEX_RADIX_SHIFT = 4;
-
-			/// <summary>
-			/// Complex data: mask to extract radix information (after shifting by <see cref="COMPLEX_RADIX_SHIFT"/>).
-			/// This give us 4 possible fixed point representations as defined below.
-			/// </summary>
-			private const Int32 COMPLEX_RADIX_MASK = 0x3;
-
-			private static String[] DIMENSION_UNIT_STRS = new String[] { "px", "dip", "sp", "pt", "in", "mm", };
-			private static String[] FRACTION_UNIT_STRS = new String[] { "%", "%p", };
-
-			private static Single MANTISSA_MULT = 1.0f / (1 << COMPLEX_MANTISSA_SHIFT);
-
-			private static Single[] RADIX_MULTS = new Single[] {
-				1.0f*MANTISSA_MULT, 1.0f/(1<<7)*MANTISSA_MULT,
-				1.0f/(1<<15)*MANTISSA_MULT, 1.0f/(1<<23)*MANTISSA_MULT
-			};
-
-			public String DataToString()
-			{
-				switch(this.ValueType)
-				{
-				case AttributeValue.NULL:
-					return "";
-				case AttributeValue.REFERENCE:
-					return "@" + value;
-				case AttributeValue.ATTRIBUTE:
-					return "?" + value;
-				case AttributeValue.FLOAT:
-					Byte[] bytes = BitConverter.GetBytes(value);
-					return BitConverter.ToSingle(bytes, 0).ToString();
-				case AttributeValue.INT_BOOLEAN:
-					return this.value != 0 ? "true" : "false";
-				case AttributeValue.INT_HEX:
-					return "0x" + this.value.ToString("X4");
-				case AttributeValue.DIMENSION:
-					Single dim = complexToFloat(value);
-					return dim.ToString() + DIMENSION_UNIT_STRS[(this.value >> COMPLEX_UNIT_SHIFT) & COMPLEX_UNIT_MASK];
-				case AttributeValue.FRACTION:
-					Single fraction = complexToFloat(value) * 100;
-					return fraction.ToString() + FRACTION_UNIT_STRS[(this.value >> COMPLEX_UNIT_SHIFT) & COMPLEX_UNIT_MASK];
-				case AttributeValue.INT_COLOR_RGB8:
-					return "#" + ToColorRGB8(this.value);
-				case AttributeValue.INT_COLOR_RGB4:
-					return "#" + ToColorRGB4(this.value);
-				case AttributeValue.INT_COLOR_ARGB8:
-					return "#" + ToColorARGB8(this.value);
-				default:
-					return this.value.ToString();
-				}
-			}
-
-			private static String ToColorRGB4(Int32 data)
-			{
-				String color = data.ToString("X");
-				String r, g, b;
-				r = color.Substring(0, 1);
-				g = color.Substring(1, 1);
-				b = color.Substring(2, 1);
-				return "0" + r + "0" + g + "0" + b;
-			}
-
-			private static String ToColorRGB8(Int32 data)
-			{
-				String color = data.ToString("X");
-				String r, g, b;
-				r = color.Substring(0, 2);
-				g = color.Substring(2, 2);
-				b = color.Substring(4, 2);
-				return r + g + b;
-			}
-
-			private static String ToColorARGB8(Int32 data)
-			{
-				String color = data.ToString("X8");
-				String a, r, g, b;
-				a = color.Substring(0, 2);
-				r = color.Substring(2, 2);
-				g = color.Substring(4, 2);
-				b = color.Substring(6, 2);
-				return a + r + g + b;
-			}
-
-			/// <summary>Retrieve the base value from a complex data integer.</summary>
-			/// <remarks>
-			/// This uses the * <see cref="COMPLEX_MANTISSA_MASK"/> and <see cref="COMPLEX_RADIX_MASK"/> fields of
-			/// the data to compute a floating point representation of the number they describe.
-			/// The units are ignored.
-			/// </remarks>
-			/// <param name="complex">A complex data value.</param>
-			/// <returns>A floating point value corresponding to the complex data.</returns>
-			private static Single complexToFloat(Int32 complex)
-				=> (complex & (COMPLEX_MANTISSA_MASK
-					<< COMPLEX_MANTISSA_SHIFT))
-					* RADIX_MULTS[(complex >> COMPLEX_RADIX_SHIFT)
-					& COMPLEX_RADIX_MASK];
+			public ResourceValueDecoder.ValueType ValueType => (ResourceValueDecoder.ValueType)(this._valueType >> 24);/*Padding +3 bytes*/
 		}
 	}
 }
